+++ Sources/MapboxNavigationCore/Map/NavigationMapViewDelegate.swift
@@ -88,8 +88,8 @@ public protocol NavigationMapViewDelegate: AnyObject, UnimplementedLogging {
         sourceIdentifier: String
     ) -> CircleLayer?
 
-    /// Asks the receiver to return a `SymbolLayer` for intermediate waypoint symbols, given an identifier and source.
-    /// This method is invoked any time intermediate waypoints are added or shown.
+    /// Asks the receiver to return a `SymbolLayer` for waypoint symbols, given an identifier and source.
+    /// This method is invoked any time waypoints are added or shown.
     /// - Parameters:
     ///   - navigationMapView: The ``NavigationMapView`` object.
     ///   - identifier: The `SymbolLayer` identifier.
@@ -102,6 +102,89 @@ public protocol NavigationMapViewDelegate: AnyObject, UnimplementedLogging {
     ) -> SymbolLayer?
 
     /// Asks the receiver to return a `FeatureCollection` that describes the geometry of waypoints.
+    ///
+    /// For example, to customize the appearance of intermediate waypoints by adding an image follow these steps:
+    ///
+    /// 1. Implement the ``NavigationMapViewDelegate/navigationMapView(_:shapeFor:legIndex:)-4osii`` method to provide a
+    /// `FeatureCollection` for waypoints.
+    /// Within this method:
+    ///     1. Add an image to the map by calling `MapboxMap.addImage(_:id:stretchX:stretchY:)` method.
+    ///     2. Iterate through `waypoints` array and create `Feature` for each waypoint.
+    ///     3. Add a key-value pair to `Feature.properties` for specifying an icon image if the waypoint is
+    ///     intermediate.
+    ///
+    /// Example:
+    ///
+    /// ```swift
+    /// func navigationMapView(
+    ///     _ navigationMapView: NavigationMapView,
+    ///     shapeFor waypoints: [Waypoint],
+    ///     legIndex: Int
+    /// ) -> FeatureCollection? {
+    ///
+    ///     let imageId = "intermediateWaypointImageId"
+    ///     if !navigationMapView.mapView.mapboxMap.imageExists(withId: imageId) {
+    ///         do {
+    ///             try navigationMapView.mapView.mapboxMap.addImage(
+    ///                 UIImage(named: "waypoint")!,
+    ///                 id: imageId,
+    ///                 stretchX: [],
+    ///                 stretchY: []
+    ///             )
+    ///         } catch {
+    ///             // Handle the error
+    ///             return nil
+    ///         }
+    ///     }
+    ///     return FeatureCollection(
+    ///         features: waypoints.enumerated().map { waypointIndex, waypoint in
+    ///             var feature = Feature(geometry: .point(Point(waypoint.coordinate)))
+    ///             var properties: [String: JSONValue] = [:]
+    ///             properties["waypointCompleted"] = .boolean(waypointIndex <= legIndex)
+    ///             properties["waypointIconImage"] = waypointIndex > 0 && waypointIndex < waypoints.count - 1
+    ///             ? .string(imageId)
+    ///             : nil
+    ///             feature.properties = properties
+    ///             return feature
+    ///         }
+    ///     )
+    /// }
+    /// ```
+    ///
+    /// 2. Implement the
+    /// ``NavigationMapViewDelegate/navigationMapView(_:waypointSymbolLayerWithIdentifier:sourceIdentifier:)-4e9yh``
+    /// method to provide a custom `SymbolLayer`.
+    ///     1. Create a `SymbolLayer`.
+    ///     2. Set `SymbolLayer.iconImage` to an expression `Exp` to retrieve the icon image name based on the
+    ///     properties defined in step 1.2.
+    ///
+    /// Example:
+    /// ```swift
+    /// func navigationMapView(
+    ///     _ navigationMapView: NavigationMapView,
+    ///     waypointSymbolLayerWithIdentifier identifier: String,
+    ///     sourceIdentifier: String
+    /// ) -> SymbolLayer? {
+    ///
+    ///     var symbolLayer = SymbolLayer(id: identifier, source: sourceIdentifier)
+    ///     let opacity = Exp(.switchCase) {
+    ///         Exp(.any) {
+    ///             Exp(.get) {
+    ///                 "waypointCompleted"
+    ///             }
+    ///         }
+    ///         0
+    ///         1
+    ///     }
+    ///     symbolLayer.iconOpacity = .expression(opacity)
+    ///     symbolLayer.iconImage = .expression(Exp(.get) { "waypointIconImage" })
+    ///     symbolLayer.iconAnchor = .constant(.bottom)
+    ///     symbolLayer.iconOffset = .constant([0, 15])
+    ///     symbolLayer.iconAllowOverlap = .constant(true)
+    ///     return symbolLayer
+    /// }
+    /// ```
+    ///
     /// - Parameters:
     ///   - navigationMapView: The ``NavigationMapView`` object.
     ///   - waypoints: The waypoints to be displayed on the map.